backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Rob Kopel <R0bk@users.noreply.github.com>" ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
llama-index = "0.10.15"
llama-index-core = "0.10.15"
llama-index-agent-openai = "0.1.5"
llama-index-embeddings-azure-openai = "0.1.6"
llama-index-llms-azure-openai = "0.1.5"
llama-index-vector-stores-qdrant = "0.1.6"
qdrant-client = "1.8.2"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.llama-index-vector-stores-milvus]
version = "^0.1.6"

[tool.poetry.dependencies.pymilvus]
version = "2.3.7"

[tool.poetry.dependencies.docx2txt]
version = "^0.8"

[tool.poetry.dependencies.llama-index-tools-google]
version = "0.1.2"

[tool.poetry.dependencies.llama-index-tools-wikipedia]
version = "0.1.2"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"